streamlit/requirements.txt
streamlit==1.52.2
requests<=2.32.5
bcrypt==4.3.0
psycopg2-binary==2.9.10
